requirements.txt
Package              Version
-------------------- -----------
pip                  22.2.2
pytest               7.1.2
PyYAML               6.0
requests             2.28.1
urllib3              1.26.12
